Plex Connect: refactor and fixes plugin (#3510)
This PR refactors the Plex Connect plugin and fixes several issues
encountered during
testing.
                                                            
Refactor
   
The monolithic player_remote.py has been split into focused modules for
better
  readability and easier contributions:                     
- server.py — HTTP server setup and shared state
  - timeline.py — Plex timeline XML generation and broadcasting
- playback.py — Playback control handlers (play, pause, seek, skip,
etc.)
- queue_commands.py — Play queue command handlers (playMedia,
createPlayQueue,
refreshPlayQueue)
- queue_sync.py — Queue synchronisation between MA and Plex
The plugin stage has also been updated from alpha to beta.
Bug fixes
                                                            
- Infinite loop on refreshPlayQueue: the refreshPlayQueue handler was
missing the
_updating_from_plex guard, causing a loop when features like Guest DJ
were used.
- ~Play queue truncated at 50 items: the default Plex API window of 50
was used when
fetching play queues. The fetch now paginates through the full queue
using
playQueueTotalCount.~ This incurs significant delays and throttling of
the playback on large queues! The queue is capped to 100 tracks.
- Queue ownership: own=True is now passed on the initial PlayQueue.get()
call so MA
takes full control of the queue.
- Shuffle not propagated to Plex: toggling shuffle from Plexamp now
triggers a
_create_plex_playqueue_from_ma call to sync the new order back.
- Double-shuffle on shuffled Plex queues: when a shuffled PlayQueue is
loaded, the
original source is now fetched via playQueueSourceURI and loaded into MA
in its
natural order. MA then applies its own shuffle and propagates the result
back to
Plex, avoiding double-shuffling an already-shuffled list.
- Stale MA shuffle corrupting unshuffled queues: the shuffle state is
now always
explicitly synced from Plex on every playMedia command, preventing a
previously
  enabled MA shuffle from silently reordering a non-shuffled queue.
