+++ music_assistant/providers/plex_connect/manifest.json
@@ -1,6 +1,6 @@
 {
   "type": "plugin",
-  "stage": "alpha",
+  "stage": "beta",
   "domain": "plex_connect",
   "name": "Plex Connect Plugin",
   "description": "Makes a Music Assistant player appear as a device in the official Plex apps.",
